third_party/WebKit/LayoutTests/http/tests/inspector/bindings/bindings-frame-attach-detach-expected.txt

Verify that UISourceCodes are added and removed as iframe gets attached and detached.


Running: dumpInitialWorkspace
Workspace: 7 uiSourceCodes.
    debugger:///VM[XXX] bindings-frame-attach-detach.html
    debugger:///VM[XXX] bindings-frame-attach-detach.html
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    http://127.0.0.1:8000/inspector/bindings/bindings-frame-attach-detach.html
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/inspector-test.js

Running: attachFrame
Workspace: 13 uiSourceCodes.
    debugger:///VM[XXX] bindings-frame-attach-detach.html
    debugger:///VM[XXX] bindings-frame-attach-detach.html
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] magic-script.js
    debugger:///VM[XXX] _test_attachFrame.js
    http://127.0.0.1:8000/inspector/bindings/bindings-frame-attach-detach.html
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/resources/magic-frame.html
    http://127.0.0.1:8000/inspector/bindings/resources/magic-script.js
    http://127.0.0.1:8000/inspector/bindings/resources/magic-style.css
    http://127.0.0.1:8000/inspector/inspector-test.js
    _test_attachFrame.js

Running: detachFrame
Workspace: 14 uiSourceCodes.
    debugger:///VM[XXX] bindings-frame-attach-detach.html
    debugger:///VM[XXX] bindings-frame-attach-detach.html
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] magic-script.js
    debugger:///VM[XXX] _test_attachFrame.js
    debugger:///VM[XXX] _test_detachFrame.js
    http://127.0.0.1:8000/inspector/bindings/bindings-frame-attach-detach.html
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/resources/magic-frame.html
    http://127.0.0.1:8000/inspector/bindings/resources/magic-script.js
    http://127.0.0.1:8000/inspector/inspector-test.js
    _test_attachFrame.js
    _test_detachFrame.js
